merge MPI messages
we often call excjs several times in a row, but on different variables.
this leads to a lot of awaiting, as the excjs is not fully async.

by allocating the variables consecutively in memory (stacking then
along the z-axis), we can send them with only a single call to excjs

In this patch, try this for in the poisson module.

NOTE: this requires fortran 90 pointer/target syntax.