Further merge MPI messages
Create an excjs4 function that exchanges 4 fields at once.
In the previous commit, we used a single allocate for a larger field, (with the original
fields pointers/subarrays) to send one big message.

Due to the non-standard bounds many of the arrays, x(2-ih:i1+ih,...),
this doesnt work for the fields in modboundary as pointers have
their lbound set to 1. So this would require changing all loops and
indices.