modbulkmicro: disable check for negative qr,Nr
This test has been ineffective for a long time due to a sign error in the sum.
Now the error is fixed, the test is spuriously triggered, especially when the total amount of qr is low(?).
For optimization and to get rid of messages we don't handle anyway, disable the test.