NH3_v06_dales-v4-3
Update to the new version of DALES: v4.3.
Also solve the thls bug from branch v02 with a clean fix.
All changes in this branch are found in the readme, changes specific to this branch (v06) are listed below.
All changes (not bugfixes) made in the code are highlighted by comments with "! START Ruben Schulte, dd-mm-yyyy"

    _v06_dales-master_dales-v4-3:
    Update data on 10-02-2021.
    Continuation based on "_v05_dales-master_perc_chem".
    This is an update from DALES v4.2 to DALES v4.3.
    All of the above changes are applied to the source code of DALES v4.3.
        modsurface.f90: Changed the bugfix for thls, mentioned in v02 of this branch, to a
        cleaner fix (line 475).