modgenstat: use 0-fields instead of m-fields for statistics
The statistics routines are run at the end of a full RK time step,
after the tstep_integrate routine. The m-fields and 0-fields are equal
at this point. But the halo cells of m-fields are not exchanged,
which matters for some calculations involving interpolated velocities.
For consistency, use 0-fields for everything.