github actions: Switch compilation test with intel compiler from ifort to ifx
By Caspar Jungbacker. Intel released a new version of the oneAPI toolkit (version 2024)
and this broke the builds in the Intel CI/CD pipelines. Fixed by removing the manual copies
of libimf, libintlc, libsvml and libirng. I also changed the compiler for NetCDF from ifort
to the newer ifx, since ifort is set to be discontinued in 2024.

Still to do:
* update the CMake file to also use ifx
* drop the Cartesius SYST option
* make the CI test use another SYST