fast_thermo0 and statistics: calculate qsat only where used
In fast_thermo0 (faster, vector-friendly thermodynamics), the qsat
value at the end is not accurate in the case the grid cell is not
saturated. Rather than doing one more function call to calculate qsat,
don't store qsat at all in the thermodynamics but calculate it as needed in
the statistics routines. The qsat field is not used anywhere
in the calculations, only for output.

After this commit the qsat 3D field can be removed.