modfftw: zero unused part of buffer on last tile before transform
The last tile in a row or column may have less data to transform.
Zero the unused data to avoid transforming uninitialized values.
The results of those should not be used anyway, but uninitialized
values may cause a floating point error in the Fourier transform
when run in debug mode.

Also remove erroneous "else" cases in the transposes.

Remove some unused imports.