update microphysics / thermodynamics for imicro_bulk3
For imicro_bulk3, the saturation adjustment in thermodynamics should be
done with respect to liquid water only. This is done by initializing
the vapor pressure table using liquid water values (for bulk3)
and ice/liquid mix values for all other cases.  Move the
initialization of those tables from modglobal to modthermodynamics,
since the initialization needs imicro and importing that into
modglobal would create a cyclic dependency.

Report an error for running bulk3 with old thermodynamics,
since that has not been updated with the liquid-only special case.