modfftw: fix zeroing of unused parts of p201
one more fix of handling the un-used part of the FFTW data.
Note: there may be several processors with no elements to transform.
Don't rely on being the last in a row or column, instead determine the
number of own elements (jonx_me, jony_me) and use those where appropriate.