GABLS case: fix namoptions, add warning for uninitialized thls
Fix GABLS flags. It uses the Mason subgrid scheme, which requires the
ldelta flag too - Mason and Deardorff cannot be used at the same time.
  ldelta = .true.  ! lmason requires ldelta
Add thls to namoptions. The surface temperature is time-dependent,
thls is still needed in namoptions to initialize the base profile.

Add a warning when initializing base profile if thls is not initialized.